requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Data Processing
openpyxl>=3.0.0
xlrd>=2.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0

# Web application
Flask>=2.0.0

# Optional: Advanced ML libraries
# optuna>=3.0.0
# scikit-optimize>=0.9.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter Notebook (for development)
jupyter>=1.0.0
ipykernel>=6.0.0

# Code Quality
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0
